requirements.txt
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Sqlalchemy==3.1.1
Flask-Wtf==1.2.1
Flask==3.0.0
PyJWT==2.8.0
black==24.4.2
email-Validator==2.1.0.post1
isort==5.13.2
python-dotenv==1.0.1
